init: Weaken property override security for the init extension
Sometimes we need to override ro.* properties by using our vendor init
extension.

Previously there was a security check which was blocking that.
To resolve the issue, we need to weaken the security check during the
execution of our vendor init extension.

This is safe because the vendor init extension gets executed as part of init
construction and it is considered a trusted system component.

Change-Id: Ia7d60686968695f1fb43be4ed58770ce10da88c5

Former-commit-id: 149e9d2e3fb48457bfdb27937040c75782a5fdb4
Former-commit-id: 29053aabd3dcc5aa3870ce05e91573ccfc478b59
Former-commit-id: 0d02e7f11fb022d344be604dad404030159e8ee5
Former-commit-id: 2f308d29b6727e1db33283b8b64f81fb4bcfdd24
Former-commit-id: ca7b3dd4686761aff1b921725e9196add68efa20